LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/initial-empty-document/iframe-nosrc-expected.txt


PASS src
PASS location.href
PASS location.assign
PASS window.open
FAIL link click assert_equals: history.length must not change after normal navigation on document loaded by iframe with no src expected 5 but got 6
FAIL form submission assert_equals: history.length increases after normal navigation from  non-initial empty document expected 7 but got 6
